requirements.txt
torch==2.2.1
transformers==4.39.1
tokenizers==0.15.2
python-dotenv==1.0.1
slack_sdk==3.27.1